Simplify and improve parse_module_list
I noticed that the parser does not accept a trailing comma:

$ cat test.pc
Name: test
Version: 1
Description: test
Requires: x11,
$ ./pkg-config --print-{requires,errors} ./test.pc
Empty package name in Requires or Conflicts in file './test.pc'

Since the parser already accepts empty modules on the left-hand side
of a comma, there is no good reason why it shouldn't do so on the
right-hand side.  Furthermore, the problem is specific to the last
comma; otherwise, the following works:

-Requires: x11,
+Requires: x11,,x11

$ pkg-config --print-{requires,errors} ./test.pc
x11
x11

To fix this, the condition for "the last module on the line"
should be changed as follows:

-  if (p != start)
+  if (state != OUTSIDE_MODULE)

Other parts of this patch simplify parsing of a module's name.
When the parser enters the IN_MODULE_NAME state, the "start" of the
name should be set to the current character.  Later, this saves a loop
in parse_module_list.

After the list is split, it is no longer possible that the name could
be empty.  This eliminates the "Empty package name" check entirely.
Also, it isn't possible that, after the list is split, the name
is followed by a comma.  Thus one instance of the MODULE_SEPARATOR()
check is now replaced with the more correct isspace().