Permit multiple Requires/Requires.private/Conflicts lines
I noticed that, for better or worse, the check for multiple
Requires lines simply doesn't work:

$ cat test.pc
Name: test
Version: 1
Description: test
Requires: x11
Requires: gl
$ ./pkg-config --print-requires ./test.pc
gl

This is because the parser works with pkg->requires_entries,
but the check is done against pkg->requires modules, which are loaded
at a later stage.

Anyway, it makes sense to permit multiple Requires lines and treat them
the same way rpm(1) and pkgconf(1) do, that is, to combine the lists.
The only implementation detail worth noting is that, since the packages
are actually prepended to the list, the reversal of the list should
be postponed.  And so now I get this:

$ ./pkg-config --print-requires ./test.pc
x11
gl