Factored verify_info() to be called before load_requires()
This restores the previous behaviour: mandatory fields like
Name and Version must be checked first, before loading Requires
and Requires.private.

In verify_info(), I remove the check for (pkg->key == NULL).
It can be seen even from the diff that pkg->key is already used
before the check, and is assumed to be non-null.  Instead,
I add an assertion at an earlier stage.