Revamped Libs and Libs.private parsing
This change simplifies parsing by virtue of all lines being fully parsed.
That is, no "ignore" instructions are now passed down to the parser -
this is simply too low and arguably wrong a level to implement the
program's logic (such as recursion).  Instead, the parser should just
do its job.  Whether to recurse upon Requires and Requires.private,
this can now be handled properly in load_requires().  This leaves
the question of whether Libs.private should be parsed, because
the parser combines both Libs and Libs.private into a single
pkg->libs list.  Well, I just let the parser parse Libs.private
into a separate list, pkg->libs_private.  Later, in verify_package(),
the lists are either combined or pkg->libs_private is discarded.
(This is only done to keep the change limited in size and scope -
maybe combining the two fields into a single list was not such a good
idea after all.  Untangling the two lists throughout the code had
better be done in a separate change, though.)

Note that parsing Libs.private into a separate list fixes the bug
that causes the following fields

Libs.private: -la
Lib: -lb

to yield '-la -lb' (instead of '-lb -la') when queried for '--static
--libs'.  Although some fields are allowed to span multiple lines,
the relative order of the fields should remain immaterial.

There can be two objections to this change: 1) more errors can
be unexpectedly triggered by the parser; 2) more work will be done
by the parser.  As for parsing errors, note that Requires and
Requires.private are the only fields which are likely to trigger
errors; this is because, unlike Libs and Libs.private, they have
a special syntax for versioned dependencies.  At the same time,
Requires and Requires.private are crucial for --cflags and even
for --exists.  When a .pc file is so badly broken that Requires
or Requires.private cannot be parsed, it is not clear if there is
a way out for pkg-config to handle such a file gracefully (such as
triggering fewer errors or, on the contrary, bailing out early).

As for performance, the impact is likely to be minuscule.  I'm going
to profile the program later and see if it needs any optimizations.
Note, however, that the existing code doesn't do selective parsing
consistently, as if for the sake of performance - e.g. the Cflags
field always gets parsed.  Indeed, selective parsing was rather used
to bolster the program's logic.

Full parsing is also crucial for the --validate mode.  Suppose a .pc
file has an undefined variable, like this:

Requires: ${foo}

Before this change, the --validate check was successful, while in
fact it shouldn't.

$ ./pkg-config.exe- --validate ./test.pc
$ ./pkg-config.exe  --validate ./test.pc
Variable 'foo' not defined in './test.pc'