Revamped dependency loading in main.c, added --disable-recursion
This is probably the last big change in this series.  It introduces the
special "tolerate missing Requires.private" mode (disabled by default)
which addresses the problem of excessive dependencies in *-devel
packages.  The comment in main.c speaks for itself, and should be
considered as part of this commit message.  This special mode was
first implemented in early 2011, and has worked well since.

To elaborate slightly on the subject of cpp.req and dependency
generators, here is one example of dependencies produced by
the distro's packaging system.

$ rpmquery --requires libcairo-devel
    libcairo = 1:1.14.4-alt1
    /usr/lib64/pkgconfig
    fontconfig-devel
    libEGL-devel
    libGL-devel
    libX11-devel
    libXrender-devel
    libfreetype-devel
    libxcb-devel
    pkgconfig(egl)
    pkgconfig(fontconfig) >= 2.2.95
    pkgconfig(freetype2) >= 9.7.3
    pkgconfig(gl)
[1] pkgconfig(libpng)
    pkgconfig(x11)
    pkgconfig(xcb) >= 1.6
    pkgconfig(xcb-render) >= 1.6
    pkgconfig(xcb-shm)
    pkgconfig(xext)
    pkgconfig(xrender) >= 0.6
[2] zlib-devel
    rpmlib(PayloadIsLzma)

The pkgconfig(*) dependencies are the Requires dependencies from
/usr/lib64/cairo*.pc.  For most of these dependencies, there is
a corresponding *-devel dependency generated by cpp.req.  For example,
the pkgconfig(x11) dependency listed in /usr/lib64/pkgconfig/cairo.pc
has the corresponding dependency on libX11-devel, due to the following
'#include' instructions:

/usr/include/cairo/cairo-xlib-xrender.h:#include <X11/Xlib.h>
/usr/include/cairo/cairo-xlib.h:#include <X11/Xlib.h>

Note, however, that the dependency [1] on pkgconfig(libpng) has no
corresponding dependency on libpng-devel.  This is because PNG support
in Cairo API does not require libpng headers.

/usr/include/cairo/cairo.h:#if CAIRO_HAS_PNG_FUNCTIONS
/usr/include/cairo/cairo.h:cairo_surface_write_to_png (cairo_surface_t
/usr/include/cairo/cairo.h:cairo_surface_write_to_png_stream (cairo_surface_t

The dependency [2] on zlib-devel is produced by yet another dependency
generator, due to the following Libs fields:

/usr/lib64/pkgconfig/cairo-pdf.pc:Libs: -lz
/usr/lib64/pkgconfig/cairo-ps.pc:Libs: -lz

The upshot is that, given the sufficiently advanced packaging system,
pkg-config internal dependencies can create more problems than they try
to solve.  In the example, the distro's packaging system "knows better"
that the pkgconfig(libpng) dependency is redundant.  It is placed there
only to appease pkg-config.  The example also suggests that Requires
dependencies in /usr/lib64/pkgconfig/cairo*.pc should be rather moved
to Requires.private (and -lz should go to Libs.private).  In other words,
it is an example of a freedesktop.org package not following its own
guidelines.

This further suggests that ALL pkg-config dependencies, both
Requires.private and Requires alike, can be treated in the "tolerate
missing" manner, in order to produce the "necessary and sufficient"
package dependencies.

The --disable-recursion option was first implemented by Dmitry V. Levin
some time in 2007.  Gee, that's ten years ago.  The option is used
in the "yet another dependency generator" to extract the original
Libs list.