Use getc_unlocked instead of getc
This makes up for the performance loss due to full parsing.

(before this change)
$ valgrind --tool=callgrind ./pkg-config- --cflags gtk+-3.0
    refs:      7,032,414

(with this change)
$ valgrind --tool=callgrind ./pkg-config  --cflags gtk+-3.0
    refs:      6,480,404

Taking into account that the function getc_unlocked is standard,
I believe it better be used in the code explicitly.  Another possibility
is redefining getc: "#define getc getc_unlocked".  This can lead to
a problem if getc_unlocked was in turn defined to be getc.